Faltaba un resize misterioso

+++ tp3/include/io.hpp
@@ -9,7 +9,7 @@ using namespace std;
 
 // Asumo escala de grises entre 0 255
 //Nota:Si uso uint8, el parser de txt piensa que son char y se va todo de mambo.
-typedef uint8_t pixel_t;
+typedef uint16_t pixel_t;
 
 typedef vector< vector<pixel_t> > frame_t;
